requirements.txt
pyvisa
https://github.com/MSLNZ/msl-equipment/archive/main.zip
ThorlabsPM100==1.2.2
icecream==2.1.2
lmfit==1.0.3
pandas==1.4
openpyxl==3.0.3
matplotlib==3.5
pyyaml==6.0
nose==1.3.7
coverage==6.3
